(refactor) Refactor to match new nomenclature
