+++ README.md
@@ -1,17 +1,12 @@
+# UgandaEMR Data Visualizer
 
-
-# UgandaEMR Reporting ESM
-
-A microfrontend for reporting in UgandaEMR built on top of OpenMRS 3.x
-
-For more information, please refer to the 
-[OpenMRS 3.x Frontend Documentation](https://openmrs.github.io/openmrs-esm-core/#/).
+A frontend module for handling data visualizations in UgandaEMR+.
 
 # Getting Started
 
 ```sh
 # Clone the repository
-git clone git@github.com:METS-Programme/esm-ugandaemr-reporting.git
+git clone git@github.com:mets-programme/esm-ugandaemr-data-visualizer.git
 
 # to install dependencies
 yarn
@@ -19,15 +14,14 @@ yarn
 # to run the dev server
 yarn start
 
-# OR to start on a specified port eg 5000
+# OR to start on a specified port e.g. 5000
 yarn start --port 5000
 ```
 
-Once it is running, a browser window
-should open with the OpenMRS 3 application. Log in and then navigate to
-`/openmrs/spa/reporting`.
+Once it is running, a browser window should launch an O3 dev server. Log in and then navigate to the home page and click the `Data visualizer` link on the left navigation menu.
 
 # Running tests
-```
+
+```sh
 yarn run test
 ```
