+++ src/components/dashboard/empty-state-illustration.component.tsx
